Update Azure AD & Exchange Online.ps1
added a check for the AIPService module and install if it is not installed, then import and connect.
added ASCII art to the modify a user menu
added a check for email encryption in the main menu, also added a function to test the encryption without actually sending an email, included a config fix for a known test error.
added a disconnect from the AIPService